refactor API code to separate files
